feat:Add digital_ocean service
